Move _stage.render() logic to shared-module
